Create source
Creates an Announcer class to announce events that are used by the Dims class which acts as a proxy for a jQuery element. This allows you to set common properties using the syntax dimsObj.left=300; More importantly, it lets you use the syntax dimsObja.left=dimsObjb.right to align the left edge of the first element with the right edge of the second.  Finally, it introduces the watch feature that triggers a callback whenever a property is modified by the dims object.